requirements.txt
# Web Scraping
beautifulsoup4==4.12.2
selenium==4.15.2
undetected-chromedriver==3.5.4
lxml==4.9.3
webdriver-manager==4.0.1

# YouTube API
google-api-python-client==2.108.0
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1

# Data Processing
pandas==2.1.3
numpy==1.26.2
python-dateutil==2.8.2

# Machine Learning
scikit-learn==1.3.2
xgboost==2.0.3
joblib==1.3.2

# Visualization
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.18.0

# Utilities
requests==2.31.0
tqdm==4.66.1
python-dotenv==1.0.0

# Jupyter
jupyter==1.0.0
notebook==7.0.6
ipykernel==6.27.1